Revert "Add details and form"
