+++ src/App.js
@@ -2,12 +2,8 @@ import Navbar from './Navbar';
 import './App.css';
 import Home from './Home';
 import Sidebar from './Sidebar';
+import BlogDetail from './BlogDetail';
 import Toggle from './Toggle';
-import { Routes, Route } from 'react-router-dom';
-import Create from './Components/create';
-import NotFound from './Components/not-found';
-import BlogDetail from './Components/BlogDetail';
-
 
 
 
@@ -31,18 +27,14 @@ function App() {
     <div className="App">
       <h1 onMouseEnter = {explore} >This is a component</h1>
       
-      <Routes>
-        <Route path='/' element={<Home/>} />
-        <Route path='/create' element={<Create/>} />
-        <Route path='*' element={<NotFound />}/>
-        <Route path= "/creators/:id" element={<BlogDetail/>}/>
-        </Routes>
-
-      
-        <Navbar/> 
+      <Home/>
+      <Navbar/> 
       <Sidebar/>
+      <BlogDetail/>
       <Toggle/>
-    
+      
+      
+      
   
     </div>
   );
