+++ src/BlogDetail.js
@@ -1,0 +1,11 @@
+const BlogDetail = () => {
+    return ( 
+        <div className="blogdetail">
+            <h1>About us</h1>
+            <p>Welcome to Webloaded - No.1 Entertainment News and Songs Hub.</p>
+
+        </div>
+     );
+}
+ 
+export default BlogDetail;
