+++ src/Home.js
@@ -1,29 +1,45 @@
 import { GiSpiderWeb } from "react-icons/gi";
 import { useState } from "react";
-import Creatorlist from "./Components/Creatorlist";
-import useFetch from "./hooks/use-fetch";
 
 const Home = () => {
 
-const[course, setCourse]= useState('open');
+    const [Creator, setCreator]= useState(
+        [
+            {id:1, author: "Tom", title: "How to work"},
+            {id:2,  author: "John", title: "Learning how to code" },
+            {id:3,  author: "Andrew", title: "Achieving your goals"},
 
-const {creator, error,loading} = useFetch('http://localhost:8000/creators')
+        ]
+    );
 
     return (  
-        <div classname="Home">
-        <h1
+        <div classname="home">
+            <h1
             style={{
                 color: "greenyellow",
         }}>Webloaded<GiSpiderWeb /></h1>
-         {loading && <div>Your item is being fetched</div>}
-        {error && <div>{error}</div>}
-       {creator && <Creatorlist creator={creator}/>}
-<button onClick={()=>setCourse('close')}>click to change course </button> 
 
-<p>{course}</p>
-</div>
+        {Creator.map(( creator,index)=>(
+            <div className="creator-preview" key={creator.id}>
+                <h2> Title is {creator.title}</h2>
+                <p>written by {creator.author}</p>
 
- );
+                </div>
+        
+
+        ))}
+
+                
+ 
+
+
+
+            </div>
+                
+                
+            
+    
+    );
 }
  
 export default Home;
