+++ src/Navbar.js
@@ -1,5 +1,4 @@
 import {useState} from "react";
-import { Link } from "react-router-dom";
 
 const Navbar = () => {
 
@@ -12,10 +11,10 @@ function handleClick(){
     return (  
         <div className="navbar">
             <h1 onClick={handleClick}>My Blog Page{count}</h1>
-            <Link to="/">home</Link><br></br>
-            <Link to="/">Back</Link><br></br>
-            <Link to="/create">Add blog</Link>
-            
+            <a href="/">home</a> <br></br>
+            <a href="/create">New blog</a><br></br>
+            <a href="/create">Back</a><br></br>
+            <a href="https://google.com">google</a><br></br>
